Removed all conditional logic and fixed formatting
