+++ src/ElCamino.AspNetCore.Identity.AzureTable/Helpers/KeyHelper.cs
@@ -7,25 +7,14 @@ using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
-#if net45
-using Microsoft.AspNet.Identity;
-#else
 using Microsoft.AspNetCore.Identity;
-#endif
 
-
-#if net45
-namespace ElCamino.AspNet.Identity.AzureTable.Helpers
-#else
 namespace ElCamino.AspNetCore.Identity.AzureTable.Helpers
-#endif
 {
     public static class KeyHelper
     {
         private static BaseKeyHelper baseHelper = new UriEncodeKeyHelper();
-#if !net45
         private static BaseKeyHelper hashHelper = new HashKeyHelper();
-#endif
 
         public static string GeneratePartitionKeyIndexByLogin(string plainLoginProvider, string plainProviderKey)
         {
@@ -62,7 +51,7 @@ namespace ElCamino.AspNetCore.Identity.AzureTable.Helpers
         {
             return baseHelper.GenerateRowKeyIdentityUserClaim(claimType, claimValue);
         }
-#if !net45
+
         public static string GenerateRowKeyIdentityRoleClaim(string claimType, string claimValue)
         {
             return hashHelper.GenerateRowKeyIdentityRoleClaim(claimType, claimValue);
@@ -73,7 +62,6 @@ namespace ElCamino.AspNetCore.Identity.AzureTable.Helpers
             return hashHelper.GenerateRowKeyIdentityUserToken(loginProvider, name);
         }
 
-#endif
         public static string GenerateRowKeyIdentityUserLogin(string loginProvider, string providerKey)
         {
             return baseHelper.GenerateRowKeyIdentityUserLogin(loginProvider, providerKey);
