+++ src/ElCamino.AspNetCore.Identity.AzureTable/IdentityAzureTableBuilderExtensions.cs
@@ -1,27 +1,18 @@
 ﻿// MIT License Copyright 2017 (c) David Melendez. All rights reserved. See License.txt in the project root for license information.
 
-#if !net45
 using System;
-#if net45
-using Microsoft.AspNet.Identity;
-using Microsoft.AspNet.Builder;
-using ElCamino.AspNet.Identity.AzureTable;
-using ElCamino.AspNet.Identity.AzureTable.Model;
-#else
 using Microsoft.AspNetCore.Identity;
 using ElCamino.AspNetCore.Identity.AzureTable.Model;
 using ElCamino.AspNetCore.Identity.AzureTable;
-#endif
 
 namespace Microsoft.Extensions.DependencyInjection
 {
-	public static class IdentityAzureTableBuilderExtensions
-	{
-
-		public static IdentityBuilder AddAzureTableStores<TContext>(this IdentityBuilder builder, Func<IdentityConfiguration> configAction)
-			where TContext : IdentityCloudContext, new()
-		{
-			builder.Services.AddSingleton<IdentityConfiguration>(new Func<IServiceProvider, IdentityConfiguration>(p=> configAction()));
+    public static class IdentityAzureTableBuilderExtensions
+    {
+        public static IdentityBuilder AddAzureTableStores<TContext>(this IdentityBuilder builder, Func<IdentityConfiguration> configAction)
+            where TContext : IdentityCloudContext, new()
+        {
+            builder.Services.AddSingleton<IdentityConfiguration>(new Func<IServiceProvider, IdentityConfiguration>(p => configAction()));
 
             Type contextType = typeof(TContext);
             Type userStoreType = typeof(UserStore<,,>).MakeGenericType(builder.UserType, builder.RoleType, contextType);
@@ -37,9 +28,9 @@ namespace Microsoft.Extensions.DependencyInjection
                 roleStoreType);
 
             return builder;
-		}
+        }
 
-		public static IdentityBuilder CreateAzureTablesIfNotExists<TContext>(this IdentityBuilder builder)
+        public static IdentityBuilder CreateAzureTablesIfNotExists<TContext>(this IdentityBuilder builder)
             where TContext : IdentityCloudContext, new()
         {
             Type contextType = typeof(TContext);
@@ -47,12 +38,10 @@ namespace Microsoft.Extensions.DependencyInjection
 
             var userStore = ActivatorUtilities.GetServiceOrCreateInstance(builder.Services.BuildServiceProvider(),
                 userStoreType) as dynamic;
-            
+
             userStore.CreateTablesIfNotExists();
 
             return builder;
-            
         }
     }
 }
-#endif
