+++ src/ElCamino.AspNetCore.Identity.AzureTable/Model/IdentityUser.cs
@@ -1,29 +1,16 @@
 ﻿// MIT License Copyright 2017 (c) David Melendez. All rights reserved. See License.txt in the project root for license information.
-#if net45
-using Microsoft.AspNet.Identity;
-using ElCamino.AspNet.Identity.AzureTable.Helpers;
-#else
-using ElCamino.AspNetCore.Identity.AzureTable.Helpers;
-#endif
-using Microsoft.WindowsAzure.Storage.Table;
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Microsoft.WindowsAzure.Storage.Table;
+using ElCamino.AspNetCore.Identity.AzureTable.Helpers;
 
-#if net45
-namespace ElCamino.AspNet.Identity.AzureTable.Model
-#else
 namespace ElCamino.AspNetCore.Identity.AzureTable.Model
-#endif
 {
-public class IdentityUser : IdentityUser<string, IdentityUserLogin, IdentityUserRole, IdentityUserClaim>
-#if net45
-		, IUser
-        , IUser<string>
-#endif
-		, IGenerateKeys
+    public class IdentityUser : IdentityUser<string, IdentityUserLogin, IdentityUserRole, IdentityUserClaim>, IGenerateKeys
     {
         public IdentityUser() { }
 
@@ -58,22 +45,13 @@ public class IdentityUser : IdentityUser<string, IdentityUserLogin, IdentityUser
 
         public override string Id
         {
-            get
-            {
-                return RowKey;
-            }
-            set
-            {
-                RowKey = value;
-            }
+            get => RowKey;
+            set => RowKey = value;
         }
 
         public override string UserName
         {
-            get
-            {
-                return base.UserName;
-            }
+            get => base.UserName;
             set
             {
                 if (!string.IsNullOrWhiteSpace(value))
@@ -85,11 +63,8 @@ public class IdentityUser : IdentityUser<string, IdentityUserLogin, IdentityUser
     }
 
     public class IdentityUser<TKey, TLogin, TRole, TClaim> : TableEntity
-#if net45
-        ,IUser<TKey>
-#endif
-		where TKey : IEquatable<TKey>
-		where TLogin : IdentityUserLogin<TKey>
+        where TKey : IEquatable<TKey>
+        where TLogin : IdentityUserLogin<TKey>
         where TRole : IdentityUserRole<TKey>
         where TClaim : IdentityUserClaim<TKey>
     {
@@ -100,31 +75,30 @@ public class IdentityUser : IdentityUser<string, IdentityUserLogin, IdentityUser
             this.Logins = new List<TLogin>(10);
         }
 
-		#region Collections
+        #region Collections
         [Microsoft.WindowsAzure.Storage.Table.IgnoreProperty]
         public ICollection<TClaim> Claims { get; private set; }
 
-		[Microsoft.WindowsAzure.Storage.Table.IgnoreProperty]
-		public ICollection<TLogin> Logins { get; private set; }
+        [Microsoft.WindowsAzure.Storage.Table.IgnoreProperty]
+        public ICollection<TLogin> Logins { get; private set; }
 
-		[Microsoft.WindowsAzure.Storage.Table.IgnoreProperty]
-		public ICollection<TRole> Roles { get; private set; }
+        [Microsoft.WindowsAzure.Storage.Table.IgnoreProperty]
+        public ICollection<TRole> Roles { get; private set; }
 
-		#endregion
-#if !net45
-		public virtual string NormalizedEmail { get; set; }
+        #endregion
 
-		public virtual string NormalizedUserName { get; set; }
+        public virtual string NormalizedEmail { get; set; }
 
-		//Not sure if this is needed.
-		public virtual string ConcurrencyStamp { get; set; } = Guid.NewGuid().ToString();
+        public virtual string NormalizedUserName { get; set; }
 
-#endif
-		public virtual int AccessFailedCount { get; set; }
+        //Not sure if this is needed.
+        public virtual string ConcurrencyStamp { get; set; } = Guid.NewGuid().ToString();
+
+        public virtual int AccessFailedCount { get; set; }
 
         public virtual string Email { get; set; }
 
-		public virtual bool EmailConfirmed { get; set; }
+        public virtual bool EmailConfirmed { get; set; }
 
         [Microsoft.WindowsAzure.Storage.Table.IgnoreProperty]
         public virtual TKey Id { get; set; }
@@ -144,7 +118,5 @@ public class IdentityUser : IdentityUser<string, IdentityUserLogin, IdentityUser
         public virtual bool TwoFactorEnabled { get; set; }
 
         public virtual string UserName { get; set; }
-
-	}
-
+    }
 }
