+++ src/ElCamino.AspNetCore.Identity.AzureTable/Model/IdentityUserClaim.cs
@@ -1,22 +1,15 @@
 ﻿// MIT License Copyright 2017 (c) David Melendez. All rights reserved. See License.txt in the project root for license information.
-using Microsoft.WindowsAzure.Storage.Table;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-
-#if net45
-using ElCamino.AspNet.Identity.AzureTable.Helpers;
-
-namespace ElCamino.AspNet.Identity.AzureTable.Model
-#else
+using Microsoft.WindowsAzure.Storage.Table;
 using ElCamino.AspNetCore.Identity.AzureTable.Helpers;
 
 namespace ElCamino.AspNetCore.Identity.AzureTable.Model
-#endif
 {
-public class IdentityUserClaim : IdentityUserClaim<string>, IGenerateKeys
+    public class IdentityUserClaim : IdentityUserClaim<string>, IGenerateKeys
     {
         public IdentityUserClaim() { }
 
@@ -42,7 +35,7 @@ public class IdentityUserClaim : IdentityUserClaim<string>, IGenerateKeys
 
         public double KeyVersion { get; set; }
 
-       [Microsoft.WindowsAzure.Storage.Table.IgnoreProperty]
+        [Microsoft.WindowsAzure.Storage.Table.IgnoreProperty]
         public override string UserId
         {
             get
@@ -65,7 +58,5 @@ public class IdentityUserClaim : IdentityUserClaim<string>, IGenerateKeys
         public string Id { get; set; }
 
         public virtual TKey UserId { get; set; }
-
     }
-
 }
