+++ src/ElCamino.AspNetCore.Identity.AzureTable/Model/IdentityUserIndex.cs
@@ -1,18 +1,15 @@
 ﻿// MIT License Copyright 2017 (c) David Melendez. All rights reserved. See License.txt in the project root for license information.
-using Microsoft.WindowsAzure.Storage.Table;
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Microsoft.WindowsAzure.Storage.Table;
 
-#if net45
-namespace ElCamino.AspNet.Identity.AzureTable.Model
-#else
 namespace ElCamino.AspNetCore.Identity.AzureTable.Model
-#endif
 {
-internal class IdentityUserIndex : TableEntity
+    internal class IdentityUserIndex : TableEntity
     {
         /// <summary>
         /// Holds the userid entity key
@@ -20,6 +17,5 @@ internal class IdentityUserIndex : TableEntity
         public string Id { get; set; }
 
         public double KeyVersion { get; set; }
-
     }
 }
