+++ src/ElCamino.AspNetCore.Identity.AzureTable/Model/IdentityUserLogin.cs
@@ -1,25 +1,18 @@
 ﻿// MIT License Copyright 2017 (c) David Melendez. All rights reserved. See License.txt in the project root for license information.
-using Microsoft.WindowsAzure.Storage.Table;
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-
-#if net45
-using ElCamino.AspNet.Identity.AzureTable.Helpers;
-
-namespace ElCamino.AspNet.Identity.AzureTable.Model
-#else
+using Microsoft.WindowsAzure.Storage.Table;
 using ElCamino.AspNetCore.Identity.AzureTable.Helpers;
 
 namespace ElCamino.AspNetCore.Identity.AzureTable.Model
-#endif
 {
-public class IdentityUserLogin : IdentityUserLogin<string>, IGenerateKeys
+    public class IdentityUserLogin : IdentityUserLogin<string>, IGenerateKeys
     {
         public IdentityUserLogin() { }
-
 
         /// <summary>
         /// Generates Row and Id keys.
@@ -63,12 +56,10 @@ public class IdentityUserLogin : IdentityUserLogin<string>, IGenerateKeys
 
         public virtual string ProviderKey { get; set; }
 
-		public virtual string ProviderDisplayName { get; set; }
+        public virtual string ProviderDisplayName { get; set; }
 
-		public virtual TKey UserId { get; set; }
+        public virtual TKey UserId { get; set; }
 
         public virtual string Id { get; set; }
-
     }
-
 }
